requirements.txt
alabaster==0.7.16
alembic==1.13.1
annotated-types==0.7.0
anyio==4.6.0
ar==0.8
astroid==3.0.3
asttokens==2.4.1
attrs==23.2.0
babel==2.16.0
black==24.2.0
blinker==1.7.0
blint==2.1.6
boltons==21.0.0
boto3==1.34.45
botocore==1.34.45
bracex==2.4
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-option-group==0.5.6
colorama==0.4.6
ConfigArgParse==1.7
coverage==7.6.4
cryptography==43.0.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dnspython==2.6.1
docker==7.0.0
docutils==0.20.1
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
face==22.0.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gevent==24.10.3
geventhttpclient==2.3.1
gitdb==4.0.11
GitPython==3.1.42
glom==22.1.0
greenlet==3.1.1
gunicorn==22.0.0
h11==0.14.0
idna==3.6
imagesize==1.4.1
immutabledict==4.1.0
importlib-metadata==4.11.3
iniconfig==2.0.0
ipython==8.24.0
isort==5.13.2
itsdangerous==2.1.2
javalang==0.13.0
jedi==0.19.1
Jinja2==3.1.4
jira==3.6.0
jmespath==1.0.0
jsonpickle==3.0.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lief==0.14.1
locust==2.32.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
milksnake==0.1.6
msgpack==1.1.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
networkx==3.3
oauthlib==3.2.2
orjson==3.10.3
packageurl-python==0.15.0
packaging==23.2
parso==0.8.4
pathspec==0.12.1
peewee==3.17.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==6.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.9.0
pylint==3.0.3
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.18.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
ruff==0.2.2
s3transfer==0.10.1
schedule==1.2.2
semgrep==1.61.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-autobuild==2024.9.19
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinx-rtd-theme==3.0.1
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.39.2
symbolic==10.2.1
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.31.0
watchfiles==0.24.0
wcmatch==8.5
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.3
zipp==3.8.0
zope.event==5.0
zope.interface==7.1.0
